agent/pyproject.toml
[tool.poetry]
name = "solana_agent"
version = "0.0.1"
description = ""
authors = ["Bevan Hunt <bevan@bevanhunt.com>"]
packages = [{include = "solana_agent"}]

[tool.poetry.dependencies]
python = "3.12.7"
fastapi = "0.115.6"
uvicorn = {version = "0.32.1", extras = ["standard"]}
gunicorn = "23.0.0"
python-dotenv = "1.0.1"
pydantic = "*"
motor = "3.6.0"
cyberchipped = "2.1.10"
sse-starlette = "2.1.3"
pyjwt = "2.10.1"

[tool.poetry.group.dev.dependencies]
pytest = "*"
pytest-asyncio = "*"
pytest-cov = "*"
pytest-mock = "*"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
